requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.33
protobuf==3.20.3
pydantic==1.10.7
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0